requirements.txt
pygame==2.1.2
PyOpenGL==3.1.7
PyOpenGL_accelerate==3.1.7
opencv-python-headless==4.8.0.76